move store initialization to the init/renderer folder
